Add style_rules to lib
