+++ lib/messages.rb
@@ -1,7 +1,7 @@
 require 'colorize'
 
 module Messages
-  DONE = 'LINTER HAS FINISHED CHECKING YOUR FILE'.bold
-  LINE = '**************************************'
+  DONE = '              LINTER HAS FINISHED CHECKING YOUR FILE'.bold
+  LINE = '              **************************************'
   NO_ERRORS = 'No errors'.green.bold
 end
